The absence of time accounting for aggregated data is taken into account. The tests are also transferred to another file for compactness
